Split TierNFT into subcomponents.
